requirements-backend.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
rich
langgraph
langchain
langchain-core
langchain-community
langchain-google-genai
langchain-huggingface
langchain-chroma
chromadb
sentence-transformers
huggingface-hub
google-genai
google-adk
google-api-python-client
google-api-core
crawl4ai
nltk
beautifulsoup4
docx2txt
unstructured[pptx]
python-docx
weasyprint
PyMuPDF
pytesseract
Pillow
prompt-toolkit